refactor(preexisting): lock uniformly across Unix and Windows
Two pieces of review feedback on the pre-existing account backend.

`local_backend.py` becomes `local_storage_backend.py`, matching the
`LocalStorageBackend` class it holds. Its sibling `s3_backend.py` still
disagrees with `S3StorageBackend`; renaming that is unrelated churn.

Locking through `fcntl.flock` directly made this mode Unix-only, because
`fcntl` has no Windows build, and it was the package's only such import
against an `OS Independent` classifier. The stdlib has no uniform
primitive -- `fcntl.flock` and `msvcrt.locking` differ in granularity
and in whether they block -- so rather than hand-roll the platform
branch, the wrapper delegates to `filelock`. That was already resolved
in `uv.lock` as a transitive dependency of datasets, huggingface-hub and
virtualenv, so declaring it directly pulls in nothing new.

`_fsync_dir` is a no-op on Windows, which cannot open a directory as a
file descriptor and offers no equivalent through `os`. The file-level
fsync before the rename still runs on both platforms, so a crash cannot
publish empty content anywhere; what Windows lacks is the directory
entry flush, leaving a rename not guaranteed durable across a host
crash.
