+++ aws_bench/resource_management/storage/local_storage_backend.py
@@ -2,6 +2,7 @@
 
 import hashlib
 import os
+import sys
 import tempfile
 from pathlib import Path
 
@@ -31,7 +32,12 @@ def _etag(data: bytes) -> str:
 
 
 def _fsync_dir(path: Path) -> None:
-    """Flush ``path``'s own directory entry, so a rename into it is durable."""
+    """Flush ``path``'s own directory entry, so a rename into it is durable.
+
+    A no-op on Windows, which cannot open a directory as a file descriptor.
+    """
+    if sys.platform == "win32":
+        return
     fd = os.open(path, os.O_RDONLY)
     try:
         os.fsync(fd)
