+++ tests/resource_management/snapshot/test_snapshot_manager.py
@@ -23,7 +23,7 @@ from aws_bench.resource_management.snapshot.models import (
     StackMetadata,
 )
 from aws_bench.resource_management.storage.exceptions import StorageConflictError
-from aws_bench.resource_management.storage.local_backend import LocalStorageBackend
+from aws_bench.resource_management.storage.local_storage_backend import LocalStorageBackend
 from aws_bench.resource_management.storage.s3_backend import S3StorageBackend
